+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support `xlsx` format
+- Add parameter to change the minimum password length in the context of
+  identifying short passwords
 
 ## [0.1.3] - 2021-11-12